gnulinwiz::functionality

Module prog_fun

Functions

check_sw_install_type

Prompts the user to choose between default or custom software installation lists.

default_sw_package

Returns the default list of software packages for installation.

handle_error

Handles errors by printing a message and exiting the program.

print_license_info

Prints the GNU/Linux Config Wizard’s license information to the console.

print_setup_status_failed

Prints a red-colored failure message to indicate an unsuccessful setup.

print_setup_status_success

Prints a green-colored success message to indicate a successful setup.

set_sw_list

Collects a list of software packages from user input for installation.

validate_root_priviliges

Validates that the program is not running with root privileges.