gnulinwiz::functionality::prog_fun

Function default_sw_package

pub fn default_sw_package() -> &'static [&'static str]

Returns a default list of software packages for installation.

This function provides a predefined list of commonly used software packages (e.g., Firefox, Vim) for the “gnulinwiz” project’s post-installation setup. It is used when the user opts for the default software installation instead of a custom list. The returned slice is static to ensure consistent access across the setup process.

Returns

A static slice of strings containing package names (e.g., ["firefox", "clang", ...]).

Example

use gnulinwiz::functionality::prog_fun::default_sw_package;
let packages = default_sw_package();
assert_eq!(packages, &["firefox", "clang", "zsh", "git", "gimp", "mpv", "curl", "vim", "rkhunter"]);

See Also

software::software_setup: Uses this list for default software installation.

check_sw_install_type: Determines whether to use this list or a custom one.