gnulinwiz::functionality::prog_fun

Function print_license_info

pub fn print_license_info()

Displays the GNU/Linux Config Wizard’s license information.

This function prints the GNU General Public License (GPL) v3 notice for the “gnulinwiz” project, including a link to the full license text. It is called at the start of the setup process to inform users of the software’s licensing terms and warranty disclaimer.

Example

use gnulinwiz::functionality::prog_fun::print_license_info;
print_license_info(); // Outputs GPL v3 notice with license link

See Also

gnu_linux_default_setup: Calls this function as part of the setup process.