gnulinwiz::functionality

Module shell

Functions

change_def_shell

Changes the default shell to Zsh for the specified user.

install_omz

Installs Oh My Zsh by downloading and executing its installation script.

install_zsh_autosuggestions

Installs the Zsh Autosuggestions plugin for the specified user’s Oh My Zsh setup.

install_zsh_syntax_highlighting

Installs the Zsh Syntax Highlighting plugin for the specified user’s Oh My Zsh setup.