gnulinwiz::functionality::shell

Function install_omz

pub fn install_omz() -> i8

Installs Oh My Zsh by downloading and executing its installation script.

Downloads the Oh My Zsh installation script using curl and pipes it to bash for execution. Handles errors during process spawning, piping, or script execution, and logs detailed output in case of failure.

Returns

0 if Oh My Zsh is installed successfully.

1 if any error occurs during the installation process.

Examples

let result = install_omz();
assert_eq!(result, 0);